refactor(cli): move helpers from ArtifactGenerator to BaseGenerator
Introduce two new helpers to be used in generators not inheriting form
`ArtifactGenerator`:

- `_isGenerationSuccessful` to detect whether all target files were
  successfully created or modified.

- `_updateIndexFile` to add a single entry to a given `index.ts` file
